requirements.txt
fastapi
uvicorn
python-dotenv
openai
pandas
pinecone-client==3.0.0
sentence-transformers
